Added dry_run option cleanup code fix windows path

+++ audiobook_renamer.py
@@ -38,6 +38,15 @@ parser.add_argument(
     ),
 )
 parser.add_argument(
+    "--dry_run",
+    action="store_true",
+    help=(
+        "Performs a dry run, which means the script will simulate the process "
+        "but will not make any actual changes to the data. This is useful for "
+        "previewing the actions that would be taken without altering anything."
+    ),
+)
+parser.add_argument(
     "--output_file",
     help=(
         "Specify the path to the log file where you want the script's output to be"
@@ -91,7 +100,7 @@ def strip_invalid_file_name_characters(variable):
 
 def naming_convention(metadata):
     jsondata = jsonHandler(
-        f"{metadata}/metadata.json",
+        os.path.join(metadata, "metadata.json"),
         mode="r",
     )
 
@@ -129,49 +138,19 @@ def recursively_list_folders(folder_path):
             yield os.path.join(root, dir)
 
 
-def move_files_and_keep_inodes(source_directory, destination_directory):
-    # Check if the destination directory exists, and create it if it doesn't.
-    if not os.path.exists(destination_directory):
-        os.makedirs(destination_directory)
-
-    # List all files in the source directory.
-    files = [
-        f
-        for f in os.listdir(source_directory)
-        if os.path.isfile(os.path.join(source_directory, f))
-    ]
-
-    for file_name in files:
-        source_file = os.path.join(source_directory, file_name)
-        destination_file = os.path.join(destination_directory, file_name)
-
-        write_info_to_file_and_console(get_inode(source_file))
-
-        # Create a hard link to the file in the destination directory.
-        os.link(source_file, destination_file)
-
-        write_info_to_file_and_console(get_inode(destination_file))
-
-        # Remove the original file from the source directory.
-        os.remove(source_file)
-    write_info_to_file_and_console(source_directory)
-    os.rmdir(source_directory)
-
-
-def get_inode(path):
-    # Get the inode number of the file or directory at the given path.
-    return os.stat(path).st_ino
-
-
 duplicate_audiobooks = []
 # Recursively list the folders in the folder
 for folder in recursively_list_folders(root_folder):
     # Clean Up metadata
     if args.remove_abs_metadata:
-        if os.path.exists(f"{folder}/metadata.abs"):
-            os.remove(f"{folder}/metadata.abs")
+        if os.path.exists(os.path.join(folder, "metadata.abs")):
+            if args.dry_run:
+                write_info_to_file_and_console(f"Will remove: {os.path.join(folder, 'metadata.abs')}")
+            else:
+                write_info_to_file_and_console(f"Will remove: {os.path.join(folder, 'metadata.abs')}")
+                os.remove(os.path.join(folder, "metadata.abs"))
     # Move and rename the Audiobook into the correct folder structure
-    if os.path.exists(f"{folder}/metadata.json"):
+    if os.path.exists(os.path.join(folder, "metadata.json")):
         parent_folder_path = os.path.dirname(folder)
         new_audio_book_path, new_audio_book_folder, only_folder_name = (
             naming_convention(folder)
@@ -180,18 +159,20 @@ for folder in recursively_list_folders(root_folder):
         new_folder = os.path.join(parent_folder_path, only_folder_name)
         if folder != new_folder:
             max_rename_count = max_rename_count - 1
-            write_info_to_file_and_console(f"New Folder: {new_folder}")
             write_info_to_file_and_console(f"Current Folder: {folder}")
+            write_info_to_file_and_console(f"New Folder: {new_folder}")
             if os.path.exists(new_folder):
                 duplicate_audiobooks.append(f"{new_folder} - {folder}")
             else:
-                # Rename the folder as unable to move it yet without breaking the inode/library item
-                os.rename(folder, new_folder)
+                if args.dry_run is False:
+                    # Rename the folder as unable to move it yet without breaking the inode/library item
+                    os.rename(folder, new_folder)
     elif is_folder_empty(folder):
         if args.remove_empty_folders:
-            if os.path.exists(f"{folder}/.DS_Store"):
-                os.remove(f"{folder}/.DS_Store")
-            os.rmdir(folder)
+            if args.dry_run is False:
+                if os.path.exists(os.path.join(folder, ".DS_Store")):
+                    os.remove(os.path.join(folder, ".DS_Store"))
+                os.rmdir(folder)
             write_info_to_file_and_console(f"Removed empty folder: {folder}")
     if max_rename_count == 0:
         write_info_to_file_and_console("Hit Max Rename Limit")
